+++ src/main/java/lol/nebula/module/ModuleManager.java
@@ -62,6 +62,7 @@ public class ModuleManager {
                 new WTap(),
 
                 new AntiHunger(),
+                new ChestBackpack(),
                 new ConsoleSpammer(),
                 new CraftCarry(),
                 new FastUse(),