+++ src/main/java/net/minecraft/network/play/server/S2EPacketCloseWindow.java
@@ -8,14 +8,14 @@ import net.minecraft.network.play.INetHandlerPlayClient;
 
 public class S2EPacketCloseWindow extends Packet
 {
-    private int field_148896_a;
+    private int windowId;
     private static final String __OBFID = "CL_00001292";
 
     public S2EPacketCloseWindow() {}
 
     public S2EPacketCloseWindow(int p_i45183_1_)
     {
-        this.field_148896_a = p_i45183_1_;
+        this.windowId = p_i45183_1_;
     }
 
     public void processPacket(INetHandlerPlayClient p_148895_1_)
@@ -28,7 +28,7 @@ public class S2EPacketCloseWindow extends Packet
      */
     public void readPacketData(PacketBuffer p_148837_1_) throws IOException
     {
-        this.field_148896_a = p_148837_1_.readUnsignedByte();
+        this.windowId = p_148837_1_.readUnsignedByte();
     }
 
     /**
@@ -36,11 +36,15 @@ public class S2EPacketCloseWindow extends Packet
      */
     public void writePacketData(PacketBuffer p_148840_1_) throws IOException
     {
-        p_148840_1_.writeByte(this.field_148896_a);
+        p_148840_1_.writeByte(this.windowId);
     }
 
     public void processPacket(INetHandler p_148833_1_)
     {
         this.processPacket((INetHandlerPlayClient)p_148833_1_);
+    }
+
+    public int getWindowId() {
+        return windowId;
     }
 }